requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
--extra-index-url https://download.pytorch.org/whl/cu118

-e file:.
    # via -r requirements.in
absl-py==2.1.0
    # via tensorboard
aiohttp==3.9.3
    # via
    #   black
    #   fsspec
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.3
    # via aiohttp
attrs==23.2.0
    # via aiohttp
black[d]==24.3.0
    # via -r requirements.in
build==1.2.1
    # via pip-tools
certifi==2024.8.30
    # via requests
cffi==1.16.0
    # via soundfile
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   black
    #   pip-tools
contourpy==1.2.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
exceptiongroup==1.2.0
    # via pytest
filelock==3.13.3
    # via
    #   huggingface-hub
    #   torch
    #   transformers
    #   triton
fonttools==4.50.0
    # via matplotlib
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2024.3.1
    # via
    #   huggingface-hub
    #   pytorch-lightning
    #   torch
grpcio==1.62.1
    # via tensorboard
huggingface-hub==0.24.7
    # via
    #   tokenizers
    #   transformers
idna==3.6
    # via
    #   requests
    #   yarl
iniconfig==2.0.0
    # via pytest
jinja2==3.1.3
    # via torch
joblib==1.3.2
    # via scikit-learn
kiwisolver==1.4.5
    # via matplotlib
lightning-utilities==0.11.2
    # via
    #   pytorch-lightning
    #   torchmetrics
markdown==3.6
    # via tensorboard
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
marshmallow==3.21.1
    # via
    #   -r requirements.in
    #   marshmallow-dataclass
marshmallow-dataclass==8.6.1
    # via -r requirements.in
matplotlib==3.8.3
    # via -r requirements.in
mpmath==1.3.0
    # via sympy
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
munch==4.0.0
    # via -r requirements.in
mypy-extensions==1.0.0
    # via
    #   black
    #   typing-inspect
networkx==3.2.1
    # via torch
numpy==1.26.4
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
    #   pandas
    #   pedalboard
    #   pytorch-lightning
    #   scikit-learn
    #   scipy
    #   tensorboard
    #   torchmetrics
    #   torchvision
    #   transformers
packaging==24.0
    # via
    #   black
    #   build
    #   huggingface-hub
    #   lightning-utilities
    #   marshmallow
    #   matplotlib
    #   pytest
    #   pytorch-lightning
    #   torchmetrics
    #   transformers
pandas==2.2.1
    # via -r requirements.in
pathspec==0.12.1
    # via black
pedalboard==0.9.14
    # via -r requirements.in
pillow==10.2.0
    # via
    #   matplotlib
    #   torchvision
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.2.0
    # via black
pluggy==1.4.0
    # via pytest
protobuf==5.26.1
    # via tensorboard
pycparser==2.21
    # via cffi
pyparsing==3.1.2
    # via matplotlib
pyproject-hooks==1.0.0
    # via
    #   build
    #   pip-tools
pytest==8.1.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytorch-lightning==2.0.8
    # via -r requirements.in
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   pytorch-lightning
    #   transformers
regex==2024.9.11
    # via transformers
requests==2.32.3
    # via
    #   huggingface-hub
    #   torchvision
    #   transformers
safetensors==0.4.5
    # via transformers
scikit-learn==1.4.1.post1
    # via -r requirements.in
scipy==1.12.0
    # via scikit-learn
six==1.16.0
    # via
    #   python-dateutil
    #   tensorboard
soundfile==0.12.1
    # via -r requirements.in
sympy==1.12
    # via torch
tensorboard==2.16.2
    # via -r requirements.in
tensorboard-data-server==0.7.2
    # via tensorboard
threadpoolctl==3.4.0
    # via scikit-learn
tokenizers==0.19.1
    # via transformers
tomli==2.0.1
    # via
    #   black
    #   build
    #   pip-tools
    #   pyproject-hooks
    #   pytest
torch==2.1.0+cu118
    # via
    #   -r requirements.in
    #   pytorch-lightning
    #   torchaudio
    #   torchmetrics
    #   torchvision
torchaudio==2.1.0+cu118
    # via -r requirements.in
torchmetrics==1.3.2
    # via pytorch-lightning
torchvision==0.16.0+cu118
    # via -r requirements.in
tqdm==4.66.2
    # via
    #   huggingface-hub
    #   pytorch-lightning
    #   transformers
transformers==4.44.2
    # via -r requirements.in
triton==2.1.0
    # via torch
typing-extensions==4.10.0
    # via
    #   black
    #   huggingface-hub
    #   lightning-utilities
    #   marshmallow-dataclass
    #   pytorch-lightning
    #   torch
    #   typing-inspect
typing-inspect==0.9.0
    # via marshmallow-dataclass
tzdata==2024.1
    # via pandas
urllib3==2.0.7
    # via
    #   -r requirements.in
    #   requests
werkzeug==3.0.1
    # via tensorboard
wheel==0.43.0
    # via pip-tools
yarl==1.9.4
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
